requirements.txt
pandas==2.2.2
numpy>=1.26.0
yfinance>=0.2.41
ta>=0.11.0
arch>=6.4
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
statsmodels>=0.14.2
nltk>=3.9
vaderSentiment>=3.3.2
transformers>=4.43.0
torch>=2.2.0
snscrape>=0.7.0
praw>=7.7.1
pytrends>=4.9.2
python-dotenv>=1.0.1
plotly>=5.22.0
streamlit>=1.36.0
tqdm>=4.66.0
pydantic>=2.8.0
